Начальная версия для ревью.
