+++ src/components/ingredients-category/ingredients-category.tsx
@@ -2,21 +2,13 @@ import { forwardRef, useMemo } from 'react';
 import { TIngredientsCategoryProps } from './type';
 import { TIngredient } from '@utils-types';
 import { IngredientsCategoryUI } from '../ui/ingredients-category';
-import { useSelector, useDispatch } from '../../services/store';
+import { useSelector } from '../../services/store';
 import { selectOrderItems } from '../../services/selectors/order-selector';
 
 export const IngredientsCategory = forwardRef<
   HTMLUListElement,
   TIngredientsCategoryProps
 >(({ title, titleRef, ingredients }, ref) => {
-  /** TODO: взять переменную из стора */
-  /*
-  const burgerConstructor = {
-    bun: {
-      _id: ''
-    },
-    ingredients: []
-    */
   const orderItems = useSelector(selectOrderItems);
   const burgerConstructor = {
     bun: orderItems.bun ? orderItems.bun : { _id: '', price: 0 },
