+++ src/components/order-card/order-card.tsx
@@ -6,16 +6,13 @@ import { TIngredient } from '@utils-types';
 import { OrderCardUI } from '../ui/order-card';
 
 import { selectAllIngredients } from '../../services/selectors/ingredients-selector';
-
-import { useSelector, useDispatch } from '../../services/store';
+import { useSelector } from '../../services/store';
 
 const maxIngredients = 6;
 
 export const OrderCard: FC<OrderCardProps> = memo(({ order }) => {
   const location = useLocation();
 
-  /** TODO: взять переменную из стора */
-  //const ingredients: TIngredient[] = [];
   const ingredients: TIngredient[] = useSelector(selectAllIngredients);
 
   const orderInfo = useMemo(() => {
