+++ src/components/orders-list/orders-list.tsx
@@ -7,8 +7,6 @@ export const OrdersList: FC<OrdersListProps> = memo(({ orders }) => {
   const orderByDate = [...orders].sort(
     (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
   );
-  console.log('OrdersList:', orders);
-  console.log('OrdersList_byDate:', orderByDate);
 
   return <OrdersListUI orderByDate={orderByDate} />;
 });
