+++ src/pages/profile/profile.tsx
@@ -6,27 +6,12 @@ import { useSelector, useDispatch } from '../../services/store';
 import { TUser } from '@utils-types';
 
 export const Profile: FC = () => {
-  /** TODO: взять переменную из стора */
-  /*
-  const user = {
-    name: useSelector(selectUserName),
-    email: useSelector(selectUserEmail)
-  };
-  */
   const storeUser = useSelector(selectGetUser);
   const user: TUser = storeUser ? storeUser : { name: '', email: '' };
   const dispatch = useDispatch();
 
   const [formValue, setFormValue] = useState({ ...user, password: '' });
-  /*
-  useEffect(() => {
-    setFormValue((prevState) => ({
-      ...prevState,
-      name: user?.name || '',
-      email: user?.email || ''
-    }));
-  }, [user]);
-*/
+
   useEffect(() => {
     setFormValue((prevState) => ({
       ...prevState,
@@ -63,7 +48,7 @@ export const Profile: FC = () => {
       [e.target.name]: e.target.value
     }));
   };
-  /**/
+
   return (
     <ProfileUI
       formValue={formValue}
@@ -73,20 +58,4 @@ export const Profile: FC = () => {
       handleInputChange={handleInputChange}
     />
   );
-  /*
-  return (
-    <ProfileUI
-      formValue={{
-        name: formValue.name ? formValue.name : '',
-        email: formValue.email ? formValue.email : '',
-        password: formValue.password
-      }}
-      isFormChanged={isFormChanged}
-      handleCancel={handleCancel}
-      handleSubmit={handleSubmit}
-      handleInputChange={handleInputChange}
-    />
-  );
-  return null;
-  */
 };
