+++ src/services/store.ts
@@ -6,11 +6,8 @@ import {
   useSelector as selectorHook
 } from 'react-redux';
 
-//const rootReducer = () => {}; // Заменить на импорт настоящего редьюсера
-
 const store = configureStore({
   reducer: rootReducer,
-  //middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(),
   devTools: process.env.NODE_ENV !== 'production'
 });
